Split utils module and add multidef implementation
